+++ Code/Caffeine/Application/Renderer/VertexBuffer.cpp
@@ -54,7 +54,12 @@ void VertexBuffer::bind(const VertexFormat &vertexFormat, const Shader &shader) 
       if(index != NOT_USED)
       {
         glEnableVertexAttribArray(index);
-        glVertexAttribPointer(index, attrib.size, attrib.type, GL_FALSE, vertexFormat.getStride(), (void*)attrib.pointer);
+        glVertexAttribPointer(index,
+                              static_cast<GLint>(attrib.size),
+                              attrib.type,
+                              GL_FALSE,
+                              vertexFormat.getStride(),
+                              (void*)attrib.pointer);
 
         GL_ERROR("Failed applying vertex format '" + attrib.name + "'")
       }
